+++ tests/Integration/MultisiteTest.php
@@ -30,6 +30,7 @@ use Automattic\LegacyRedirector\Domain\SourceUrl;
  * @uses \Automattic\LegacyRedirector\Domain\DestinationPostId
  * @uses \Automattic\LegacyRedirector\Domain\DestinationUrl
  * @uses \Automattic\LegacyRedirector\Domain\Redirect
+ * @uses \Automattic\LegacyRedirector\Domain\RedirectStatus
  * @uses \Automattic\LegacyRedirector\Domain\SourceUrl
  */
 final class MultisiteTest extends TestCase {